requirements.txt
streamlit 
pymongo 
psycopg2-binary 
faker 
pandas 
plotly 
python-dotenv